backend/requirements.txt
Django==4.2.16
djangorestframework==3.15.2
djoser==2.3.0
djangorestframework-simplejwt==5.3.1
drf-extra-fields==3.7.0
django-filter==24.3
Pillow==11.0.0
psycopg2-binary==2.9.10
gunicorn==23.0.0
python-dotenv==1.0.1
